+++ app/src/main/java/xyz/kewiany/menusy/data/source/remote/ProductService.kt
@@ -29,7 +29,7 @@ class ProductService @Inject constructor() : ProductApi {
                         id = "p20",
                         name = "tomato soup",
                         description = "italian tomatoes, basiland a touch of fennel",
-                        price = 10.90f,
+                        price = 10.90.toBigDecimal(),
                         menuId = "m0"
                     )
                 )
@@ -41,7 +41,7 @@ class ProductService @Inject constructor() : ProductApi {
                                 id = "p0",
                                 name = "chicken wings",
                                 description = "crispy chicken wings, louisiana hot sauce, buttermilk blue cheese ranch dressing, carrot, celery",
-                                price = 31.90f,
+                                price = 31.90.toBigDecimal(),
                                 menuId = "m1",
                                 categoryId = "c0"
                             ),
@@ -49,7 +49,7 @@ class ProductService @Inject constructor() : ProductApi {
                                 id = "p1",
                                 name = "salmon tatar",
                                 description = "salmon, soy sauce, chives, cucumber, mustard, purple potato chips",
-                                price = 36.90f,
+                                price = 36.90.toBigDecimal(),
                                 menuId = "m1",
                                 categoryId = "c0"
                             ),
@@ -61,7 +61,7 @@ class ProductService @Inject constructor() : ProductApi {
                                 id = "p2",
                                 name = "apple cake",
                                 description = "warm apple cake with shortcrust pastry, vanilla ice cream",
-                                price = 22.90f,
+                                price = 22.90.toBigDecimal(),
                                 menuId = "m1",
                                 categoryId = "c1"
                             )
@@ -73,7 +73,7 @@ class ProductService @Inject constructor() : ProductApi {
                                 id = "p3",
                                 name = "tea",
                                 description = "black, earl grey, green, jasmine, chamomile, mint",
-                                price = 14.00f,
+                                price = 14.00.toBigDecimal(),
                                 menuId = "m1",
                                 categoryId = "c2"
                             ),
@@ -81,7 +81,7 @@ class ProductService @Inject constructor() : ProductApi {
                                 id = "p4",
                                 name = "almond frappe",
                                 description = "black or with milk",
-                                price = 18.00f,
+                                price = 18.00.toBigDecimal(),
                                 menuId = "m1",
                                 categoryId = "c2"
                             )
@@ -93,7 +93,7 @@ class ProductService @Inject constructor() : ProductApi {
                                 id = "p5",
                                 name = "toma juice",
                                 description = "apple, orange, grapefruit, tomato",
-                                price = 9.00f,
+                                price = 9.00.toBigDecimal(),
                                 menuId = "m1",
                                 categoryId = "c3"
                             ),
@@ -101,7 +101,7 @@ class ProductService @Inject constructor() : ProductApi {
                                 id = "p6",
                                 name = "cisowianka",
                                 description = "still, sparkling",
-                                price = 19.00f,
+                                price = 19.00.toBigDecimal(),
                                 menuId = "m1",
                                 categoryId = "c3"
                             )
@@ -126,14 +126,14 @@ class ProductService @Inject constructor() : ProductApi {
             id = "p20",
             name = "tomato soup",
             description = "italian tomatoes, basiland a touch of fennel",
-            price = 10.90f,
+            price = 10.90.toBigDecimal(),
             menuId = "m0"
         ),
         Product(
             id = "p0",
             name = "chicken wings",
             description = "crispy chicken wings, louisiana hot sauce, buttermilk blue cheese ranch dressing, carrot, celery",
-            price = 31.90f,
+            price = 31.90.toBigDecimal(),
             menuId = "m1",
             categoryId = "c0"
         ),
@@ -141,7 +141,7 @@ class ProductService @Inject constructor() : ProductApi {
             id = "p1",
             name = "salmon tatar",
             description = "salmon, soy sauce, chives, cucumber, mustard, purple potato chips",
-            price = 36.90f,
+            price = 36.90.toBigDecimal(),
             menuId = "m1",
             categoryId = "c0"
         ),
@@ -149,7 +149,7 @@ class ProductService @Inject constructor() : ProductApi {
             id = "p2",
             name = "apple cake",
             description = "warm apple cake with shortcrust pastry, vanilla ice cream",
-            price = 22.90f,
+            price = 22.90.toBigDecimal(),
             menuId = "m1",
             categoryId = "c1"
         ),
@@ -157,7 +157,7 @@ class ProductService @Inject constructor() : ProductApi {
             id = "p3",
             name = "tea",
             description = "black, earl grey, green, jasmine, chamomile, mint",
-            price = 14.00f,
+            price = 14.00.toBigDecimal(),
             menuId = "m1",
             categoryId = "c2"
         ),
@@ -165,7 +165,7 @@ class ProductService @Inject constructor() : ProductApi {
             id = "p4",
             name = "almond frappe",
             description = "black or with milk",
-            price = 18.00f,
+            price = 18.00.toBigDecimal(),
             menuId = "m1",
             categoryId = "c2"
         ),
@@ -173,7 +173,7 @@ class ProductService @Inject constructor() : ProductApi {
             id = "p5",
             name = "toma juice",
             description = "apple, orange, grapefruit, tomato",
-            price = 9.00f,
+            price = 9.00.toBigDecimal(),
             menuId = "m1",
             categoryId = "c3"
         ),
@@ -181,7 +181,7 @@ class ProductService @Inject constructor() : ProductApi {
             id = "p6",
             name = "cisowianka",
             description = "still, sparkling",
-            price = 19.00f,
+            price = 19.00.toBigDecimal(),
             menuId = "m1",
             categoryId = "c3"
         )
